# Improve performance of nanmean/nansum on large arrays