--- a/src/ArrayStats/nanmean.jl
+++ b/src/ArrayStats/nanmean.jl
@@ -76,110 +76,116 @@ end
 # Generate customized set of loops for a given ndims and a vector
 # `static_dims` of dimensions to reduce over
 function staticdim_nanmean_quote(static_dims::Vector{Int}, N::Int)
-  M = length(static_dims)
-  # `static_dims` now contains every dim we're taking the mean over.
-  Bᵥ = Expr(:call, :view, :B)
-  reduct_inds = Int[]
-  nonreduct_inds = Int[]
-  # Firstly, build our expressions for indexing each array
-  Aind = :(A[])
-  Bind = :(Bᵥ[])
-  inds = Vector{Symbol}(undef, N)
-  for n ∈ 1:N
-    ind = Symbol(:i_,n)
-    inds[n] = ind
-    push!(Aind.args, ind)
-    if n ∈ static_dims
-      push!(reduct_inds, n)
-      push!(Bᵥ.args, :(firstindex(B,$n)))
-    else
-      push!(nonreduct_inds, n)
-      push!(Bᵥ.args, :)
-      push!(Bind.args, ind)
+    M = length(static_dims)
+    # `static_dims` now contains every dim we're taking the mean over.
+    Bᵥ = Expr(:call, :view, :B)
+    reduct_inds = Int[]
+    nonreduct_inds = Int[]
+
+    # Firstly, build our expressions for indexing each array
+    Aind = :(A[])
+    Bind = :(Bᵥ[])
+    inds = Vector{Symbol}(undef, N)
+    for n ∈ 1:N
+        ind = Symbol(:i_,n)
+        inds[n] = ind
+        push!(Aind.args, ind)
+        if n ∈ static_dims
+            push!(reduct_inds, n)
+            push!(Bᵥ.args, :(firstindex(B,$n)))
+        else
+            push!(nonreduct_inds, n)
+            push!(Bᵥ.args, :)
+            push!(Bind.args, ind)
+        end
     end
-  end
-  firstn = first(nonreduct_inds)
-  # Secondly, build up our set of loops
-  block = Expr(:block)
-  loops = Expr(:for, :($(inds[firstn]) = indices((A,B),$firstn)), block)
-  if length(nonreduct_inds) > 1
-    for n ∈ @view(nonreduct_inds[2:end])
-      newblock = Expr(:block)
-      push!(block.args, Expr(:for, :($(inds[n]) = indices((A,B),$n)), newblock))
-      block = newblock
+
+    # Secondly, build up our set of loops
+    firstn = first(nonreduct_inds)
+    block = Expr(:block)
+    loops = Expr(:for, :($(inds[firstn]) = indices((A,B),$firstn)), block)
+    if length(nonreduct_inds) > 1
+        for n ∈ @view(nonreduct_inds[2:end])
+            newblock = Expr(:block)
+            push!(block.args, Expr(:for, :($(inds[n]) = indices((A,B),$n)), newblock))
+            block = newblock
+        end
     end
-  end
-  rblock = block
-  # Push more things here if you want them at the beginning of the reduction loop
-  push!(rblock.args, :(n = 0))
-  push!(rblock.args, :(Σ = ∅))
-  # Build the reduction loop
-  for n ∈ reduct_inds
-    newblock = Expr(:block)
-    if n==last(reduct_inds)
-      push!(block.args, Expr(:macrocall, Symbol("@simd"), :ivdep, Expr(:for, :($(inds[n]) = axes(A,$n)), newblock)))
-    else
-      push!(block.args, Expr(:for, :($(inds[n]) = axes(A,$n)), newblock))
+
+    # Push more things here if you want them at the beginning of the reduction loop
+    rblock = block
+    push!(rblock.args, :(n = 0))
+    push!(rblock.args, :(Σ = ∅))
+    # Build the reduction loop
+    for n ∈ reduct_inds
+        newblock = Expr(:block)
+        if n==last(reduct_inds)
+            push!(block.args, Expr(:macrocall, Symbol("@simd"), :ivdep, Expr(:for, :($(inds[n]) = axes(A,$n)), newblock)))
+        else
+            push!(block.args, Expr(:for, :($(inds[n]) = axes(A,$n)), newblock))
+        end
+        block = newblock
     end
-    block = newblock
-  end
-  # Push more things here if you want them in the innermost loop
-  push!(block.args, :(Aᵢ = $Aind))
-  push!(block.args, :(notnan = Aᵢ==Aᵢ))
-  push!(block.args, :(n += notnan))
-  push!(block.args, :(Σ += ifelse(notnan, Aᵢ, ∅)))
-  # Push more things here if you want them at the end of the reduction loop
-  push!(rblock.args, :($Bind = Σ * inv(n)))
-  # Put it all together
-  quote
-    ∅ = zero(eltype(B))
-    Bᵥ = $Bᵥ
-    @inbounds $loops
-    return B
-  end
+
+    # Push more things here if you want them in the innermost loop
+    push!(block.args, :(Aᵢ = $Aind))
+    push!(block.args, :(notnan = Aᵢ==Aᵢ))
+    push!(block.args, :(n += notnan))
+    push!(block.args, :(Σ += ifelse(notnan, Aᵢ, ∅)))
+
+    # Push more things here if you want them at the end of the reduction loop
+    push!(rblock.args, :($Bind = Σ * inv(n)))
+
+    # Put it all together
+    quote
+        ∅ = zero(eltype(B))
+        Bᵥ = $Bᵥ
+        @inbounds $loops
+        return B
+    end
 end
 
 # Turn non-static integers in `dims` tuple into `StaticInt`s
 # so we can construct `static_dims` vector within @generated code
 function branches_nanmean_quote(N::Int, M::Int, D)
-  static_dims = Int[]
-  for m ∈ 1:M
-    param = D.parameters[m]
-    if param <: StaticInt
-      new_dim = _dim(param)::Int
-      @assert new_dim ∉ static_dims
-      push!(static_dims, new_dim)
-    else
-      t = Expr(:tuple)
-      for n ∈ static_dims
-        push!(t.args, :(StaticInt{$n}()))
-      end
-      q = Expr(:block, :(dimm = dims[$m]))
-      qold = q
-      ifsym = :if
-      for n ∈ 1:N
-        n ∈ static_dims && continue
-        tc = copy(t)
-        push!(tc.args, :(StaticInt{$n}()))
-        qnew = Expr(ifsym, :(dimm == $n), :(return _nanmean!(B, A, $tc)))
-        for r ∈ m+1:M
-          push!(tc.args, :(dims[$r]))
+    static_dims = Int[]
+    for m ∈ 1:M
+        param = D.parameters[m]
+        if param <: StaticInt
+            new_dim = _dim(param)::Int
+            @assert new_dim ∉ static_dims
+            push!(static_dims, new_dim)
+        else
+            t = Expr(:tuple)
+            for n ∈ static_dims
+                push!(t.args, :(StaticInt{$n}()))
+            end
+            q = Expr(:block, :(dimm = dims[$m]))
+            qold = q
+            ifsym = :if
+            for n ∈ 1:N
+                n ∈ static_dims && continue
+                tc = copy(t)
+                push!(tc.args, :(StaticInt{$n}()))
+                qnew = Expr(ifsym, :(dimm == $n), :(return _nanmean!(B, A, $tc)))
+                for r ∈ m+1:M
+                    push!(tc.args, :(dims[$r]))
+                end
+                push!(qold.args, qnew)
+                qold = qnew
+                ifsym = :elseif
+            end
+            push!(qold.args, Expr(:block, :(throw("Dimension `$dimm` not found."))))
+            return q
         end
-        push!(qold.args, qnew)
-        qold = qnew
-        ifsym = :elseif
-      end
-      push!(qold.args, Expr(:block, :(throw("Dimension `$dimm` not found."))))
-      return q
     end
-  end
-  staticdim_nanmean_quote(static_dims, N)
+    staticdim_nanmean_quote(static_dims, N)
 end
 
 # Efficient @generated in-place mean
 @generated function _nanmean!(B::AbstractArray{Tₒ,N}, A::AbstractArray{T,N}, dims::D) where {Tₒ,T,N,M,D<:Tuple{Vararg{IntOrStaticInt,M}}}
-  N == M && return :(B[1] = _nanmean(A, :); B)
-  branches_nanmean_quote(N, M, D)
+    N == M && return :(B[1] = _nanmean(A, :); B)
+    branches_nanmean_quote(N, M, D)
 end
 
 ## ---