… Project.toml
… README.md
… benchmark/benchmarks.jl
… test/Project.toml
… test/runtests.jl
… test/testArrayStats.jl