This implements some of the things described in #54, namely implementations for nanmean/nansum using patterns from Base. The changes to the loop ordering in 1e8c7c8 could be applied to the other functions too but I was lazy :sloth: I think it'd be good if we could reduce the amount of code duplication in the future.

The changes unfortunately ended up being a lot more invasive than I hoped because of the extra `size_threshold` argument (naming is up for debate). I added some benchmarks as well (smaller is better for all columns):

<details>
<summary>AMD Epyc server</summary>

|                            | dirty            | main             | dirty / main  |
|:---------------------------|:----------------:|:----------------:|:-------------:|
| nanmean/big array, dim 1   | 2.92 ± 0.044 ms  | 2.98 ± 0.049 ms  | 0.979 ± 0.022 |
| nanmean/big array, dim 2   | 11.1 ± 0.037 ms  | 13.1 ± 0.8 ms    | 0.844 ± 0.051 |
| nanmean/big array, dim 3   | 14.4 ± 3.5 ms    | 18.2 ± 0.91 ms   | 0.792 ± 0.2   |
| nanmean/small array, dim 1 | 0.59 ± 0.011 μs  | 0.6 ± 0.01 μs    | 0.983 ± 0.025 |
| nanmean/small array, dim 2 | 0.97 ± 0.01 μs   | 0.98 ± 0.01 μs   | 0.99 ± 0.014  |
| nanmean/small array, dim 3 | 0.77 ± 0.01 μs   | 0.87 ± 0.002 μs  | 0.885 ± 0.012 |
| nanmean/small vector       | 31 ± 10 ns       | 30 ± 1 ns        | 1.03 ± 0.34   |
| nansum/big array, dim 1    | 2.75 ± 0.062 ms  | 2.83 ± 0.05 ms   | 0.972 ± 0.028 |
| nansum/big array, dim 2    | 2.82 ± 0.045 ms  | 10 ± 0.44 ms     | 0.282 ± 0.013 |
| nansum/big array, dim 3    | 3.26 ± 0.51 ms   | 18.6 ± 0.7 ms    | 0.175 ± 0.028 |
| nansum/small array, dim 1  | 0.589 ± 0.01 μs  | 0.54 ± 0.011 μs  | 1.09 ± 0.029  |
| nansum/small array, dim 2  | 0.551 ± 0.01 μs  | 0.611 ± 0.02 μs  | 0.902 ± 0.034 |
| nansum/small array, dim 3  | 0.53 ± 0.05 μs   | 0.72 ± 0.019 μs  | 0.736 ± 0.072 |
| nansum/small vector        | 30 ± 0 ns        | 30 ± 0 ns        | 1 ± 0         |
| time_to_load               | 0.189 ± 0.0022 s | 0.189 ± 0.0032 s | 1 ± 0.021     |

</details>

<details>
<summary>Intel server</summary>

|                            | dirty            | main               | dirty / main   |
|:---------------------------|:----------------:|:------------------:|:--------------:|
| nanmean/big array, dim 1   | 9.5 ± 0.6 ms     | 8.79 ± 0.51 ms     | 1.08 ± 0.093   |
| nanmean/big array, dim 2   | 14.2 ± 0.26 ms   | 0.0332 ± 8.9e-05 s | 0.429 ± 0.0078 |
| nanmean/big array, dim 3   | 21.4 ± 4.3 ms    | 0.0362 ± 0.0019 s  | 0.591 ± 0.12   |
| nanmean/small array, dim 1 | 1.27 ± 0.053 μs  | 1.3 ± 0.019 μs     | 0.975 ± 0.043  |
| nanmean/small array, dim 2 | 1.62 ± 0.017 μs  | 1.68 ± 0.017 μs    | 0.964 ± 0.014  |
| nanmean/small array, dim 3 | 1.5 ± 0.033 μs   | 1.44 ± 0.015 μs    | 1.04 ± 0.025   |
| nanmean/small vector       | 0.049 ± 0.003 μs | 0.051 ± 0.004 μs   | 0.961 ± 0.096  |
| nansum/big array, dim 1    | 8.4 ± 0.54 ms    | 8.02 ± 0.6 ms      | 1.05 ± 0.1     |
| nansum/big array, dim 2    | 8.35 ± 0.23 ms   | 25.6 ± 0.086 ms    | 0.326 ± 0.009  |
| nansum/big array, dim 3    | 13.8 ± 1.1 ms    | 24 ± 1.8 ms        | 0.573 ± 0.063  |
| nansum/small array, dim 1  | 0.911 ± 0.031 μs | 1.05 ± 0.006 μs    | 0.871 ± 0.03   |
| nansum/small array, dim 2  | 1.01 ± 0.071 μs  | 1.22 ± 0.019 μs    | 0.825 ± 0.06   |
| nansum/small array, dim 3  | 1.05 ± 0.12 μs   | 0.965 ± 0.007 μs   | 1.09 ± 0.13    |
| nansum/small vector        | 0.046 ± 0.001 μs | 0.044 ± 0.004 μs   | 1.05 ± 0.098   |
| time_to_load               | 0.258 ± 0.015 s  | 0.264 ± 0.0094 s   | 0.976 ± 0.068  |

</details>

<details>
<summary>Intel laptop</summary>

Nansum improvements are weirdly massive here for some reason.

|                            | dirty            | main               | dirty / main   |
|:---------------------------|:----------------:|:------------------:|:--------------:|
| nanmean/big array, dim 1   | 4.74 ± 0.051 ms  | 4.83 ± 0.043 ms    | 0.981 ± 0.014  |
| nanmean/big array, dim 2   | 10.2 ± 0.045 ms  | 0.0493 ± 0.00095 s | 0.208 ± 0.0041 |
| nanmean/big array, dim 3   | 14.3 ± 3.1 ms    | 0.0564 ± 0.002 s   | 0.254 ± 0.056  |
| nanmean/small array, dim 1 | 0.736 ± 0.081 μs | 0.768 ± 0.057 μs   | 0.958 ± 0.13   |
| nanmean/small array, dim 2 | 1.48 ± 0.14 μs   | 1.6 ± 0.21 μs      | 0.922 ± 0.15   |
| nanmean/small array, dim 3 | 1.64 ± 0.16 μs   | 1.69 ± 0.17 μs     | 0.965 ± 0.13   |
| nanmean/small vector       | 0.04 ± 0.003 μs  | 0.046 ± 0.002 μs   | 0.87 ± 0.075   |
| nansum/big array, dim 1    | 4.32 ± 0.07 ms   | 4.4 ± 0.061 ms     | 0.981 ± 0.021  |
| nansum/big array, dim 2    | 4.59 ± 0.07 ms   | 0.045 ± 0.00011 s  | 0.102 ± 0.0016 |
| nansum/big array, dim 3    | 9.23 ± 0.7 ms    | 0.0572 ± 0.0014 s  | 0.161 ± 0.013  |
| nansum/small array, dim 1  | 0.858 ± 0.074 μs | 0.959 ± 0.11 μs    | 0.895 ± 0.13   |
| nansum/small array, dim 2  | 0.937 ± 0.075 μs | 1.15 ± 0.11 μs     | 0.813 ± 0.1    |
| nansum/small array, dim 3  | 0.867 ± 0.073 μs | 1.26 ± 0.12 μs     | 0.687 ± 0.086  |
| nansum/small vector        | 0.04 ± 0.001 μs  | 0.043 ± 0.001 μs   | 0.93 ± 0.032   |
| time_to_load               | 0.154 ± 0.0079 s | 0.16 ± 0.0086 s    | 0.962 ± 0.071  |

</details>

I've tried to keep all the commits atomic so I'd recommend reviewing them one-by-one.

EDIT: two issues in CI:
- Something is very wrong on 32bit, maybe an overflow somewhere.
- ~~On nightly there's a nan in the output :grimacing:~~ Dumb mistake, fixed in 4fca8ce.