@JamesWrigley 2025-05-23
I've made a few changes in the latest version:
- Fixed the 32bit issue by initializing the output array first.
- Significantly simplified the implementation by removing the `size_threshold` argument from most of the functions, since they already allow taking a `mean` argument I figure people who want something other than the defaults can compute it themselves.

Updated benchmarks (they look much the same as before):

<details>
<summary>AMD Epyc server</summary>

|                            | dirty            | main             | dirty / main   |
|:---------------------------|:----------------:|:----------------:|:--------------:|
| nanmean/big array, dim 1   | 2.95 ± 0.048 ms  | 2.97 ± 0.035 ms  | 0.993 ± 0.02   |
| nanmean/big array, dim 2   | 10.8 ± 0.023 ms  | 13.4 ± 0.028 ms  | 0.806 ± 0.0024 |
| nanmean/big array, dim 3   | 13.5 ± 0.87 ms   | 18 ± 0.74 ms     | 0.749 ± 0.058  |
| nanmean/small array, dim 1 | 0.57 ± 0.01 μs   | 0.58 ± 0.019 μs  | 0.983 ± 0.037  |
| nanmean/small array, dim 2 | 0.99 ± 0.01 μs   | 0.98 ± 0.01 μs   | 1.01 ± 0.015   |
| nanmean/small array, dim 3 | 0.72 ± 0.01 μs   | 0.89 ± 0.011 μs  | 0.809 ± 0.015  |
| nanmean/small vector       | 30 ± 10 ns       | 30 ± 10 ns       | 1 ± 0.47       |
| nansum/big array, dim 1    | 2.79 ± 0.032 ms  | 2.82 ± 0.042 ms  | 0.988 ± 0.019  |
| nansum/big array, dim 2    | 2.85 ± 0.025 ms  | 10.6 ± 0.6 ms    | 0.27 ± 0.016   |
| nansum/big array, dim 3    | 3.49 ± 0.34 ms   | 17 ± 0.66 ms     | 0.206 ± 0.022  |
| nansum/small array, dim 1  | 0.53 ± 0.01 μs   | 0.54 ± 0.011 μs  | 0.981 ± 0.027  |
| nansum/small array, dim 2  | 0.55 ± 0.03 μs   | 0.61 ± 0.02 μs   | 0.902 ± 0.057  |
| nansum/small array, dim 3  | 0.531 ± 0.061 μs | 0.71 ± 0.01 μs   | 0.748 ± 0.087  |
| nansum/small vector        | 30 ± 0 ns        | 30 ± 0 ns        | 1 ± 0          |
| time_to_load               | 0.194 ± 0.0034 s | 0.198 ± 0.0065 s | 0.981 ± 0.037  |


</details>

<details>
<summary>Intel server</summary>

|                            | dirty            | main              | dirty / main   |
|:---------------------------|:----------------:|:-----------------:|:--------------:|
| nanmean/big array, dim 1   | 9.43 ± 0.59 ms   | 8.82 ± 0.32 ms    | 1.07 ± 0.077   |
| nanmean/big array, dim 2   | 13.7 ± 0.18 ms   | 0.0326 ± 0.0003 s | 0.418 ± 0.0068 |
| nanmean/big array, dim 3   | 17.3 ± 1.6 ms    | 0.0343 ± 0.0014 s | 0.504 ± 0.05   |
| nanmean/small array, dim 1 | 1.16 ± 0.022 μs  | 1.19 ± 0.016 μs   | 0.973 ± 0.023  |
| nanmean/small array, dim 2 | 1.5 ± 0.038 μs   | 1.53 ± 0.02 μs    | 0.98 ± 0.028   |
| nanmean/small array, dim 3 | 1.34 ± 0.01 μs   | 1.31 ± 0.014 μs   | 1.03 ± 0.013   |
| nanmean/small vector       | 0.047 ± 0.001 μs | 0.046 ± 0.001 μs  | 1.02 ± 0.031   |
| nansum/big array, dim 1    | 8.19 ± 0.77 ms   | 7.95 ± 0.61 ms    | 1.03 ± 0.12    |
| nansum/big array, dim 2    | 8.34 ± 0.18 ms   | 24.5 ± 0.051 ms   | 0.341 ± 0.0076 |
| nansum/big array, dim 3    | 13.7 ± 0.82 ms   | 22.7 ± 1.7 ms     | 0.603 ± 0.058  |
| nansum/small array, dim 1  | 0.827 ± 0.026 μs | 0.796 ± 0.008 μs  | 1.04 ± 0.034   |
| nansum/small array, dim 2  | 0.938 ± 0.083 μs | 1.27 ± 0.015 μs   | 0.742 ± 0.066  |
| nansum/small array, dim 3  | 0.857 ± 0.12 μs  | 1.05 ± 0.055 μs   | 0.815 ± 0.12   |
| nansum/small vector        | 0.042 ± 0.001 μs | 0.042 ± 0.001 μs  | 1 ± 0.034      |
| time_to_load               | 0.228 ± 0.001 s  | 0.217 ± 0.019 s   | 1.05 ± 0.093   |

</details>

<details>
<summary>Intel laptop</summary>

|                            | dirty            | main               | dirty / main     |
|:---------------------------|:----------------:|:------------------:|:----------------:|
| nanmean/big array, dim 1   | 4.69 ± 0.043 ms  | 4.75 ± 0.041 ms    | 0.987 ± 0.012    |
| nanmean/big array, dim 2   | 10.2 ± 0.045 ms  | 0.047 ± 0.00088 s  | 0.218 ± 0.0042   |
| nanmean/big array, dim 3   | 13.5 ± 2.5 ms    | 0.0582 ± 0.0015 s  | 0.231 ± 0.043    |
| nanmean/small array, dim 1 | 0.733 ± 0.12 μs  | 0.731 ± 0.059 μs   | 1 ± 0.18         |
| nanmean/small array, dim 2 | 1.49 ± 0.5 μs    | 1.51 ± 0.16 μs     | 0.985 ± 0.35     |
| nanmean/small array, dim 3 | 1.62 ± 0.2 μs    | 1.66 ± 0.16 μs     | 0.981 ± 0.15     |
| nanmean/small vector       | 0.044 ± 0.002 μs | 0.046 ± 0.001 μs   | 0.957 ± 0.048    |
| nansum/big array, dim 1    | 4.29 ± 0.088 ms  | 4.33 ± 0.044 ms    | 0.991 ± 0.023    |
| nansum/big array, dim 2    | 4.49 ± 0.041 ms  | 0.0451 ± 8.2e-05 s | 0.0995 ± 0.00093 |
| nansum/big array, dim 3    | 9.08 ± 0.83 ms   | 0.0572 ± 0.0014 s  | 0.159 ± 0.015    |
| nansum/small array, dim 1  | 0.844 ± 0.081 μs | 0.844 ± 0.086 μs   | 1 ± 0.14         |
| nansum/small array, dim 2  | 0.918 ± 0.1 μs   | 1.14 ± 0.11 μs     | 0.807 ± 0.12     |
| nansum/small array, dim 3  | 0.858 ± 0.12 μs  | 1.24 ± 0.11 μs     | 0.69 ± 0.11      |
| nansum/small vector        | 0.043 ± 0.002 μs | 0.043 ± 0.002 μs   | 1 ± 0.066        |
| time_to_load               | 0.154 ± 0.0011 s | 0.152 ± 0.00098 s  | 1.01 ± 0.0099    |

</details>