@brenhinkeller 2025-05-25
Looks like a clear win on performance!